possible fix for site URL issues
It looks good. Since you're publishing from the `docs` folder, you will likely need to provide `url` and `baseurl` variables in your `_config.yml` file. These are suggestions. You'll have to load it and see if they work. If not, it may require some more configuration. If they work, this should fix the current issue, which is probably the references to CSS files and any other static files are broken, as are the links between pages (like to your post, etc).